Rename gammapy/cube/basic_cube.py -> gammapy/cube/counts.py
