+++ gammapy/cube/__init__.py
@@ -4,7 +4,7 @@ Sky cubes (3-dimensional: energy, lon, lat).
 """
 from .core import *
 from .models import *
-from .basic_cube import *
+from .counts import *
 from .fit import *
 from .new import *
 from .psf_kernel import *
