+++ gammapy/cube/counts.py
